client/rest refactoring
This heavy-weight refactoring sprang from an attempt
of removing a duplicate PostProcessResponse() function
located in client/keys/utils.go that turned out to be
undoable without introducing import cycles.
Thus I moved all x/ modules-agnostic HTTP/REST basic
request and response types and convenience functions
into a new cosmos-sdk/types/rest standalone package.
The cosmos-sdk/client/rest package retains the txs
handling functions that depend on CLIContext and
TxBuilder provided by client/context and
x/auth/client/txbuilder respectively.
