+++ client/rest/rest.go
@@ -1,9 +1,7 @@
 package rest
 
 import (
-	"fmt"
 	"net/http"
-	"strconv"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/context"
@@ -11,90 +9,10 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/keyerror"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	. "github.com/cosmos/cosmos-sdk/types/rest"
 	"github.com/cosmos/cosmos-sdk/x/auth"
 	authtxb "github.com/cosmos/cosmos-sdk/x/auth/client/txbuilder"
 )
-
-//-----------------------------------------------------------------------------
-// Basic HTTP utilities
-
-// ErrorResponse defines the attributes of a JSON error response.
-type ErrorResponse struct {
-	Code    int    `json:"code,omitempty"`
-	Message string `json:"message"`
-}
-
-// NewErrorResponse creates a new ErrorResponse instance.
-func NewErrorResponse(code int, msg string) ErrorResponse {
-	return ErrorResponse{Code: code, Message: msg}
-}
-
-// WriteErrorResponse prepares and writes a HTTP error
-// given a status code and an error message.
-func WriteErrorResponse(w http.ResponseWriter, status int, err string) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	w.Write(codec.Cdc.MustMarshalJSON(NewErrorResponse(0, err)))
-}
-
-// WriteSimulationResponse prepares and writes an HTTP
-// response for transactions simulations.
-func WriteSimulationResponse(w http.ResponseWriter, cdc *codec.Codec, gas uint64) {
-	gasEst := GasEstimateResponse{GasEstimate: gas}
-	resp, err := cdc.MarshalJSON(gasEst)
-	if err != nil {
-		WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	w.Write(resp)
-}
-
-// ParseInt64OrReturnBadRequest converts s to a int64 value.
-func ParseInt64OrReturnBadRequest(w http.ResponseWriter, s string) (n int64, ok bool) {
-	var err error
-
-	n, err = strconv.ParseInt(s, 10, 64)
-	if err != nil {
-		err := fmt.Errorf("'%s' is not a valid int64", s)
-		WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-		return n, false
-	}
-
-	return n, true
-}
-
-// ParseUint64OrReturnBadRequest converts s to a uint64 value.
-func ParseUint64OrReturnBadRequest(w http.ResponseWriter, s string) (n uint64, ok bool) {
-	var err error
-
-	n, err = strconv.ParseUint(s, 10, 64)
-	if err != nil {
-		err := fmt.Errorf("'%s' is not a valid uint64", s)
-		WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-		return n, false
-	}
-
-	return n, true
-}
-
-// ParseFloat64OrReturnBadRequest converts s to a float64 value. It returns a
-// default value, defaultIfEmpty, if the string is empty.
-func ParseFloat64OrReturnBadRequest(w http.ResponseWriter, s string, defaultIfEmpty float64) (n float64, ok bool) {
-	if len(s) == 0 {
-		return defaultIfEmpty, true
-	}
-
-	n, err := strconv.ParseFloat(s, 64)
-	if err != nil {
-		WriteErrorResponse(w, http.StatusBadRequest, err.Error())
-		return n, false
-	}
-
-	return n, true
-}
 
 //-----------------------------------------------------------------------------
 // Building / Sending utilities
@@ -177,30 +95,6 @@ func CompleteAndBroadcastTxREST(
 	}
 
 	PostProcessResponse(w, cdc, res, cliCtx.Indent)
-}
-
-// PostProcessResponse performs post processing for a REST response.
-func PostProcessResponse(w http.ResponseWriter, cdc *codec.Codec, response interface{}, indent bool) {
-	var output []byte
-
-	switch response.(type) {
-	default:
-		var err error
-		if indent {
-			output, err = cdc.MarshalJSONIndent(response, "", "  ")
-		} else {
-			output, err = cdc.MarshalJSON(response)
-		}
-		if err != nil {
-			WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
-			return
-		}
-	case []byte:
-		output = response.([]byte)
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.Write(output)
 }
 
 // WriteGenerateStdTxResponse writes response for the generate only mode.
