+++ client/rpc/codec.go
@@ -1,7 +1,7 @@
 package rpc
 
 import (
-	amino "github.com/tendermint/go-amino"
+	"github.com/tendermint/go-amino"
 	ctypes "github.com/tendermint/tendermint/rpc/core/types"
 )
