+++ client/rpc/status.go
@@ -5,15 +5,14 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/cosmos/cosmos-sdk/client/rest"
-
 	"github.com/spf13/cobra"
-
 	"github.com/spf13/viper"
+
 	ctypes "github.com/tendermint/tendermint/rpc/core/types"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/context"
+	"github.com/cosmos/cosmos-sdk/types/rest"
 )
 
 // StatusCommand returns the status of the network
