+++ x/auth/client/rest/encode.go
@@ -5,8 +5,9 @@ import (
 	"net/http"
 
 	"github.com/cosmos/cosmos-sdk/client/context"
-	"github.com/cosmos/cosmos-sdk/client/rest"
 	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/cosmos/cosmos-sdk/types/rest"
+
 	"github.com/cosmos/cosmos-sdk/x/auth"
 )
