+++ x/slashing/client/rest/rest.go
@@ -1,11 +1,11 @@
 package rest
 
 import (
+	"github.com/gorilla/mux"
+
 	"github.com/cosmos/cosmos-sdk/client/context"
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/crypto/keys"
-
-	"github.com/gorilla/mux"
 )
 
 // RegisterRoutes registers staking-related REST handlers to a router
