+++ x/staking/client/rest/utils.go
@@ -4,18 +4,17 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/cosmos/cosmos-sdk/client/rest"
-
 	"github.com/gorilla/mux"
+
+	rpcclient "github.com/tendermint/tendermint/rpc/client"
 
 	"github.com/cosmos/cosmos-sdk/client/context"
 	"github.com/cosmos/cosmos-sdk/client/tx"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/rest"
 	"github.com/cosmos/cosmos-sdk/x/staking"
 	"github.com/cosmos/cosmos-sdk/x/staking/tags"
-
-	rpcclient "github.com/tendermint/tendermint/rpc/client"
 )
 
 // contains checks if the a given query contains one of the tx types
